[APP-2865] Turn feature-mmp module into specific apkfy module
