+++ feature-apkfy/src/main/AndroidManifest.xml
