+++ feature-apkfy/src/main/java/cm/aptoide/pt/feature_apkfy/repository/ApkfyRepository.kt
@@ -1,0 +1,8 @@
+package cm.aptoide.pt.feature_apkfy.repository
+
+import cm.aptoide.pt.feature_apkfy.domain.ApkfyModel
+
+interface ApkfyRepository {
+
+  suspend fun getApkfy(): ApkfyModel
+}
