--- feature-mmp/src/main/java/cm/aptoide/pt/feature_mmp/apkfy/repository/ApkfyRepository.kt
@@ -1,8 +1,0 @@
-package cm.aptoide.pt.feature_mmp.apkfy.repository
-
-import cm.aptoide.pt.feature_mmp.apkfy.domain.ApkfyModel
-
-interface ApkfyRepository {
-
-  suspend fun getApkfy(): ApkfyModel
-}
